Replace old payment gateways class logic with the react one
